Add files via upload
This project contains 4 different classes. The main class is UnitConverter which creates a window displaying a list of buttons that can be pressed based on what unit of measurement the user wishes to convert. The length, mass, and volume classes are classes that when their respective button from the main window are pressed are responsible for creating a separate window and performing unit conversion calculations and displaying them.